requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
gspread>=6.0.0
google-auth>=2.28.0
google-api-python-client>=2.120.0
# Legacy (mantido para compatibilidade, pode remover depois)
flask>=3.0.0
gunicorn>=21.2.0
